// src/services/rest-api/customer-service/model/financial-info-response-data-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Customer Service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.267
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { FinancialInfoResponse } from './financial-info-response';

/**
 * 
 * @export
 * @interface FinancialInfoResponseDataResponse
 */
export interface FinancialInfoResponseDataResponse {
    /**
     * 
     * @type {number}
     * @memberof FinancialInfoResponseDataResponse
     */
    'status'?: number;
    /**
     * 
     * @type {{ [key: string]: Array<string>; }}
     * @memberof FinancialInfoResponseDataResponse
     */
    'errors'?: { [key: string]: Array<string>; } | null;
    /**
     * 
     * @type {string}
     * @memberof FinancialInfoResponseDataResponse
     */
    'message'?: string | null;
    /**
     * 
     * @type {FinancialInfoResponse}
     * @memberof FinancialInfoResponseDataResponse
     */
    'data'?: FinancialInfoResponse | null;
}